Fix pin level holding during deep sleep.
The camera power regulators were not shut down properly because the
enable pin level locked at the wrong level. Also flash was probably not
working because flash pin level was locked during init. (I clearly
didn't read the documentation of the hold function).

After reading the documentation I still don't understand how this
works... but this seems to work as I want it to.