Increase minimum sleep time, to prevent camera issues.
The whole power down and power up/reset procedure of the camera is not
according to spec. This due to the fact that the reset line and PWRDN
line are not controllable from the MCU, and we only power down the 1.2
and 2.8 Volt power rails. When powering back up again to fast the camera
can end up in a incorrect state and not work.