Rework putting camera to sleep
There are now two options to put the camera to sleep in deep sleep. One
is the method by shutting down the low voltage regulators. But this is
not suggested. Other option is to modify the board so that the PWDN pin
of the camera is connected to the MCU.

Also reordered some of the defines and includes at the top of the file
and added an option WITH_FLASH.