Remove NTP and GNSS support
Focus development on low power time-lapse camera. The NTP and GNSS
options are not compatible with low power operation. So removed them.

Currently there is no way to configure the time. This is planned to be
fixed by the set-up mode, once implemented.

The old firmware with NTP and GNSS code is still available in the
branch 'old_ntp_gnss'.